LayoutTests/media/media-controls-drag-timeline-set-controls-property-expected.txt

Test that dragging the timeline slider after setting video.controls=true causes seeks.


EVENT(playing)
EVENT(pause)
EVENT(seeked)
EVENT(seeked)
EVENT(seeked)
END OF TEST
